webview: extract injected Hermes WebUI scripts
