+++ app/src/test/java/com/hermeswebui/android/webui/HermesWebUiScriptsTest.kt
@@ -1,0 +1,38 @@
+package com.hermeswebui.android.webui
+
+import com.google.common.truth.Truth.assertThat
+import org.junit.Test
+
+class HermesWebUiScriptsTest {
+    @Test
+    fun `app settings script preserves folded navigation selectors`() {
+        val script = HermesWebUiScripts.appSettingsEntryScript
+
+        assertThat(script).contains("width < 799")
+        assertThat(script).contains("button.nav-tab.has-tooltip--bottom[data-tooltip=\"Settings\"]")
+        assertThat(script).contains(".mobile-nav button[data-tooltip=\"Settings\"]")
+        assertThat(script).contains(".bottom-nav button[data-tooltip=\"Settings\"]")
+        assertThat(script).contains("findCompactSettingsAnchor() || findAnchorByKind('settings') || findAnchorByKind('help')")
+    }
+
+    @Test
+    fun `app settings script routes to native application settings deep link`() {
+        val script = HermesWebUiScripts.appSettingsEntryScript
+
+        assertThat(script).contains("var appSettingsHref = 'hermes://app/settings';")
+        assertThat(script).contains("window.location.href = appSettingsHref;")
+        assertThat(script).contains("Application Settings")
+    }
+
+    @Test
+    fun `notification bridge builder injects bridge name and permission`() {
+        val script = HermesWebUiScripts.buildNotificationBridgeScript(
+            bridgeName = "HermesAndroidNotifications",
+            initialPermission = "granted"
+        )
+
+        assertThat(script).contains("var bridgeName = \"HermesAndroidNotifications\";")
+        assertThat(script).contains("var initialPermission = \"granted\";")
+        assertThat(script).contains("window.__hermesAndroidSetNotificationPermission")
+    }
+}
